Fix getDataDir in Paths_* module
In the case of alex the result of `getDataDir` in the generated
`Paths_*` module would point to the absolute path `/data/`. The reason
was that the `%{base_dir}` that was substituted in by Bazel into
`paths-template.hs` was `/data/`, which when combined with `</>` would
override the left-hand side. In case of `alex` `ctx.label.package` is
empty which causes `base_dir` to become `/data`. Replacing string
concatenation by `paths.join` resolves the issue.