Define *-cbits on empty libraries
If a Cabal library does not expose any modules, Hazel takes a short-cut
and defines an empty cc_library target instead of building a
haskell_library target. This mechanism had not been updated to also
define a *-cbits target as is required since commit 7ffc6ec. This commit
changes that.

This issue comes up on the bytestring-builder package. The
bytestring-builder package defaults to an empty library, since recent
versions of bytestring contain the bytestring-builder package.